Improve network switch handling on Send requests received.
