+++ app/components/Nav/App/index.js
@@ -153,18 +153,24 @@ const App = ({ userLoggedIn }) => {
 		[]
 	);
 
+	const frequentRpcList = useSelector((state) => state.engine.backgroundState.PreferencesController.frequentRpcList);
+
 	useEffect(() => {
 		SharedDeeplinkManager.init({
-			navigate: (routeName, opts) => {
-				const params = { name: routeName, params: opts };
-				navigator.current?.dispatch?.(CommonActions.navigate(params));
+			navigation: {
+				navigate: (routeName, opts) => {
+					const params = { name: routeName, params: opts };
+					navigator.current?.dispatch?.(CommonActions.navigate(params));
+				},
 			},
+			frequentRpcList,
+			dispatch,
 		});
 
 		unsubscribeFromBranch.current = branchSubscriber.subscribe();
 
 		return () => unsubscribeFromBranch.current?.();
-	}, [branchSubscriber]);
+	}, [branchSubscriber, frequentRpcList, dispatch]);
 
 	useEffect(() => {
 		const initAnalytics = async () => {
