+++ app/components/UI/ReceiveRequest/index.js
@@ -237,7 +237,7 @@ class ReceiveRequest extends PureComponent {
 									}}
 								>
 									<QRCode
-										value={`ethereum:${this.props.selectedAddress}`}
+										value={`ethereum:${this.props.selectedAddress}@${this.props.network}`}
 										size={Dimensions.get('window').width / 2}
 									/>
 								</TouchableOpacity>
